% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peBS.R
\name{peBS}
\alias{peBS}
\title{Probability estimator of BS density.}
\usage{
peBS(x, y, h)
}
\arguments{
\item{x}{estimate points}

\item{y}{samples}

\item{h}{bandwidth}
}
\value{
estimated density at x
}
\description{
This function is used to compute the estimated Bart Simpson density .
}
\examples{
\dontrun{
x<-c(1,2,3)
y<-rnorm(20)
h<-0.5
peBS(x,y,h)
}
}
