% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBS.R
\name{rBS}
\alias{rBS}
\title{Random points from BS density.}
\usage{
rBS(n)
}
\arguments{
\item{n}{sample size}
}
\value{
random sample from BS destiny
}
\description{
This function is used to generate random samples from Bart Simpson density.
}
\examples{
\dontrun{
n<-1000
y<-rBS(1000)
}
}
