% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwMC}
\alias{rwMC}
\title{Random walk Metropolis sampler using Rcpp.}
\usage{
rwMC(x0, sigma, N)
}
\arguments{
\item{x0}{initial point}

\item{sigma}{normal distribution variance}

\item{n}{sample size}
}
\value{
x a random sample of size \code{n}

k the number of iterations
}
\description{
Implement a random walk Metropolis sampler for generating the standard Laplace distribution
}
\examples{
\dontrun{
x0<-25
sigma<-0.05
n<-2000
rwMC(x0,sigma,n)
}
}
